backend/requirements.txt
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
DateTime==5.5
dnspython==2.6.1
email_validator==2.2.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
geographiclib==2.0
geopy==2.4.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyJWT==2.9.0
pymongo==4.3.2
python-dotenv==1.0.0
pytz==2024.2
requests==2.32.3
six==1.16.0
tenacity==9.0.0
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
zope.interface==7.1.1
